% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Sparse Vector Mean}
\usage{
\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object}
}
\description{
Computes the mean of a sparse numeric vector
}
