% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Sparse vector norm}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments}
}
\description{
Computes the Euclidean norm of a sparse numeric vector
}
\section{Methods (by class)}{
\itemize{
\item \code{norm(sparse_numeric)}: Euclidean norm

}}
