% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Two Sparse Numeric Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{sparse_numeric objects}

\item{...}{extra arguments}
}
\description{
Generates an Rplot which shows points at which two vectors are non-zero, and their values.
}
