% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Print out sparse vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{The sparse numeric vector}
}
\description{
Print out sparse vector
}
