% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add Two Sparse Vectors}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}}

\item{...}{Additional arguments}
}
\description{
Add Two Sparse Vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_add(x = sparse_numeric, y = sparse_numeric)}: Add two sparse vectors

}}
