% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_div}
\alias{sparse_div}
\alias{sparse_div,sparse_numeric,sparse_numeric-method}
\title{Elementwise divide two vectors}
\usage{
sparse_div(x, y, ...)

\S4method{sparse_div}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}}

\item{...}{Additional arguments}
}
\description{
Elementwise divide two vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_div(x = sparse_numeric, y = sparse_numeric)}: Elementwise division

}}
