% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Elementwise multiply two vectors}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}}

\item{...}{Additional arguments}
}
\description{
Elementwise multiply two vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_mult(x = sparse_numeric, y = sparse_numeric)}: Elementwise multiply

}}
