% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector Class}
\description{
An S4 class for representing sparse numeric vectors.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector giving positions of the non-zero entries.}

\item{\code{length}}{Integer giving total vector length.}
}}

