% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract Two Sparse Vectors}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}}

\item{...}{Additional arguments}
}
\description{
Subtract Two Sparse Vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_sub(x = sparse_numeric, y = sparse_numeric)}: Subtract two sparse vectors

}}
