% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_validity}
\alias{sparse_validity}
\title{Validity Check for sparse_numeric}
\arguments{
\item{object}{A \code{sparse_numeric} object}
}
\description{
Ensures that positions are within the allowed range and that `length` is valid.
}
