% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a Sparse Numeric Vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments}
}
\description{
Z-score standardizes a sparse numeric vector
}
\section{Methods (by class)}{
\itemize{
\item \code{standardize(sparse_numeric)}: Vector standardization

}}
